refactor: hub 도메인 패키지 위치 변경
