+++ hub/src/main/java/com/zgzg/hub/application/hub/res/PageHubsResDTO.java
@@ -1,4 +1,5 @@
-package com.zgzg.hub.application.res;
+package com.zgzg.hub.application.hub.res;
+
 
 import com.zgzg.hub.domain.entity.Hub;
 import java.util.UUID;
@@ -6,18 +7,19 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.data.domain.Page;
 
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CreateHubResDTO {
+public class PageHubsResDTO {
 
-  private HubDTO hubDTO;
+  private Page<HubDTO> hubs;
 
-  public static CreateHubResDTO from(Hub hub) {
-    return CreateHubResDTO.builder()
-        .hubDTO(HubDTO.from(hub))
+  public static PageHubsResDTO from(Page<Hub> hubs) {
+    return PageHubsResDTO.builder()
+        .hubs(hubs.map(HubDTO::from))
         .build();
   }
