+++ hub/src/main/java/com/zgzg/hub/presentation/hub/req/CreateHubReqDTO.java
@@ -1,4 +1,4 @@
-package com.zgzg.hub.presentation.req;
+package com.zgzg.hub.presentation.hub.req;
 
 import com.zgzg.hub.domain.entity.Hub;
 import jakarta.validation.Valid;
